kalazio/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
cycler==0.10.0
Django==3.2.13
django-adminlte3==0.1.6
django-cleanup==5.2.0
django-cors-headers==3.10.1
django-jalali==4.3.0
django-taggit==1.5.1
django-tinymce==3.3.0
djangorestframework==3.12.4
idna==3.2
jdatetime==3.6.2
kavenegar==1.1.2
kiwisolver==1.3.1
matplotlib==3.4.3
numpy==1.21.2
Pillow==9.0.1
pyparsing==2.4.7
pycryptodome==3.14.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
psycopg2
flake8
black
gunicorn
